+++ server.js
@@ -1,5 +1,3 @@
-const express = require('express')
-const bodyParser = require('body-parser')
 const { ApolloServer } = require('apollo-server')
 const mongoose = require('mongoose')
 
@@ -11,7 +9,7 @@ const server = new ApolloServer(graphqlSchema)
 mongoose
 	.connect(`mongodb://localhost:27017/${process.env.MONGO_DB}`, { useNewUrlParser: true })
 	.then(() => {
-		server.listen(process.env.PORT).then(({ url }) => {
+		server.listen(process.env.PORT || 4000).then(({ url }) => {
 			console.log(`🚀 Server ready at ${url}`);
 		  })
 	})